backend-python/requirements.txt
fastapi==0.104.0
jsonref==1.1.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
pyodbc==5.1.0
uvicorn[standard]==0.23.2
uuid6==2024.1.12
aiohttp==3.9.3
SQLAlchemy==2.0.28
python-jose==3.3.0
